requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Django==1.11.21
djangorestframework==3.9.4
kombu==4.6.1
lxml==4.3.4
psycopg2-binary==2.8.2
pytz==2019.1
redis==3.2.1
vine==1.3.0
